+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+- Add renovate bot ([1146](https://github.com/opendevstack/ods-quickstarters/pull/1146))
 - Update Quickstarter Tests to Support New Framework Test Capabilities ([#1144](https://github.com/opendevstack/ods-quickstarters/pull/1144))
 ### Added